requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
google-genai==1.53.0
pypdf==6.4.1
